Split out pep425 code, add license
